Refactor desktop client and operator modules
