+++ crates/defra-agent-desktop/src/client/mutations/mod.rs
@@ -1,0 +1,13 @@
+mod chat;
+mod graphql;
+mod operator;
+mod peers;
+
+pub use chat::{
+    create_conversation, retry_request, submit_request, CreatedConversation, SubmittedRequest,
+};
+pub use operator::{
+    run_scheduled_task_now, upsert_agent_behavior, upsert_inference_backend,
+    upsert_inference_profile, upsert_scheduled_task, upsert_tool_selection,
+};
+pub use peers::PeerMutationResult;
