+++ crates/defra-agent-desktop/src/client/mutations/peers.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct PeerMutationResult {
+    pub peer_id: String,
+    pub label: String,
+    pub addr: String,
+    pub connected: bool,
+    pub warning: Option<String>,
+}
