+++ crates/defra-agent-desktop/src/views/operator/drafts.rs
@@ -1,0 +1,450 @@
+use crate::client::ClientStore;
+use crate::state::{
+    BackendDraft, BehaviorDraft, InferenceProfileDraft, OperatorDraft, OperatorSection,
+    ScheduledTaskDraft, ToolSelectionDraft,
+};
+
+use super::recent_failures::recent_failure_summaries;
+use super::request_timeline::request_timeline_summaries;
+use super::shared::{bool_word, scheduled_task_is_due, scheduled_task_next_run_label};
+use super::EntitySummary;
+
+pub(super) fn entity_summaries(
+    store: &ClientStore,
+    section: OperatorSection,
+    selected_agent_did: Option<&str>,
+) -> Vec<EntitySummary> {
+    match section {
+        OperatorSection::Behaviors => {
+            let mut rows: Vec<_> = store
+                .behaviors
+                .iter()
+                .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                .collect();
+            rows.sort_by(|left, right| left.behavior_id.cmp(&right.behavior_id));
+            rows.into_iter()
+                .map(|row| EntitySummary {
+                    id: row.behavior_id.clone(),
+                    title: row
+                        .display_name
+                        .clone()
+                        .unwrap_or_else(|| row.behavior_id.clone()),
+                    meta: format!(
+                        "{}  model {}  backend {}",
+                        if row.enabled == Some(false) {
+                            "disabled"
+                        } else {
+                            "enabled"
+                        },
+                        row.model_name.as_deref().unwrap_or("unbound"),
+                        row.backend_id.as_deref().unwrap_or("unbound"),
+                    ),
+                })
+                .collect()
+        }
+        OperatorSection::Backends => {
+            let backend_ids = backend_ids_for_agent(store, selected_agent_did);
+            let mut rows = store
+                .inference_backends
+                .iter()
+                .filter(|row| backend_ids.contains(&row.backend_id.as_str()))
+                .collect::<Vec<_>>();
+            rows.sort_by(|left, right| left.backend_id.cmp(&right.backend_id));
+            rows.into_iter()
+                .map(|row| EntitySummary {
+                    id: row.backend_id.clone(),
+                    title: row.name.clone().unwrap_or_else(|| row.backend_id.clone()),
+                    meta: format!(
+                        "{}  probe {}  models {}",
+                        row.provider_kind.as_deref().unwrap_or("provider"),
+                        row.probe_status.as_deref().unwrap_or("unknown"),
+                        row.models.len(),
+                    ),
+                })
+                .collect()
+        }
+        OperatorSection::ToolSelections => {
+            let mut rows: Vec<_> = store
+                .tool_selections
+                .iter()
+                .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                .collect();
+            rows.sort_by(|left, right| left.selection_id.cmp(&right.selection_id));
+            rows.into_iter()
+                .map(|row| EntitySummary {
+                    id: row.selection_id.clone(),
+                    title: row
+                        .display_name
+                        .clone()
+                        .unwrap_or_else(|| row.selection_id.clone()),
+                    meta: format!(
+                        "file:{} bash:{} meta:{}",
+                        bool_word(row.enable_file_tools),
+                        bool_word(row.enable_bash),
+                        bool_word(row.enable_meta_tools),
+                    ),
+                })
+                .collect()
+        }
+        OperatorSection::InferenceProfiles => {
+            let profile_ids = inference_profile_ids_for_agent(store, selected_agent_did);
+            let mut rows = store
+                .inference_profiles
+                .iter()
+                .filter(|row| profile_ids.contains(&row.profile_id.as_str()))
+                .collect::<Vec<_>>();
+            rows.sort_by(|left, right| left.profile_id.cmp(&right.profile_id));
+            rows.into_iter()
+                .map(|row| EntitySummary {
+                    id: row.profile_id.clone(),
+                    title: row
+                        .display_name
+                        .clone()
+                        .unwrap_or_else(|| row.profile_id.clone()),
+                    meta: format!(
+                        "ctx {}  out {}  temp {}",
+                        row.context_window
+                            .map(|value| value.to_string())
+                            .unwrap_or_else(|| "na".to_string()),
+                        row.max_output_tokens
+                            .map(|value| value.to_string())
+                            .unwrap_or_else(|| "na".to_string()),
+                        row.temperature
+                            .map(|value| value.to_string())
+                            .unwrap_or_else(|| "na".to_string()),
+                    ),
+                })
+                .collect()
+        }
+        OperatorSection::ScheduledTasks => {
+            let mut rows: Vec<_> = store
+                .scheduled_tasks
+                .iter()
+                .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                .collect();
+            rows.sort_by(|left, right| {
+                left.next_run_at
+                    .cmp(&right.next_run_at)
+                    .then_with(|| left.task_id.cmp(&right.task_id))
+            });
+            rows.into_iter()
+                .map(|row| EntitySummary {
+                    id: row.task_id.clone(),
+                    title: row.name.clone().unwrap_or_else(|| row.task_id.clone()),
+                    meta: format!(
+                        "{}  every {}s  next {}  runs {}",
+                        if row.enabled == Some(false) {
+                            "disabled"
+                        } else if scheduled_task_is_due(row) {
+                            "due"
+                        } else {
+                            "armed"
+                        },
+                        row.interval_secs
+                            .map(|value| value.to_string())
+                            .unwrap_or_else(|| "na".to_string()),
+                        scheduled_task_next_run_label(row),
+                        row.run_count
+                            .map(|value| value.to_string())
+                            .unwrap_or_else(|| "0".to_string()),
+                    ),
+                })
+                .collect()
+        }
+        OperatorSection::RequestTimeline => request_timeline_summaries(store, selected_agent_did),
+        OperatorSection::RecentFailures => recent_failure_summaries(store, selected_agent_did),
+        _ => Vec::new(),
+    }
+}
+
+fn backend_ids_for_agent<'a>(
+    store: &'a ClientStore,
+    selected_agent_did: Option<&str>,
+) -> Vec<&'a str> {
+    store
+        .behaviors
+        .iter()
+        .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+        .filter_map(|row| row.backend_id.as_deref())
+        .collect()
+}
+
+fn inference_profile_ids_for_agent<'a>(
+    store: &'a ClientStore,
+    selected_agent_did: Option<&str>,
+) -> Vec<&'a str> {
+    store
+        .behaviors
+        .iter()
+        .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+        .filter_map(|row| row.inference_profile_id.as_deref())
+        .collect()
+}
+
+pub(super) fn filter_entity_summaries(
+    entries: Vec<EntitySummary>,
+    filter: &str,
+) -> Vec<EntitySummary> {
+    let filter = filter.trim().to_lowercase();
+    if filter.is_empty() {
+        return entries;
+    }
+
+    entries
+        .into_iter()
+        .filter(|entry| {
+            [entry.id.as_str(), entry.title.as_str(), entry.meta.as_str()]
+                .into_iter()
+                .any(|field| field.to_lowercase().contains(&filter))
+        })
+        .collect()
+}
+
+pub(super) fn section_meta(
+    store: &ClientStore,
+    section: OperatorSection,
+    selected_agent_did: Option<&str>,
+) -> (&'static str, String) {
+    match section {
+        OperatorSection::Runtime => (
+            "Runtime",
+            selected_agent_did
+                .and_then(|agent_did| store.latest_runtime(agent_did))
+                .and_then(|runtime| runtime.process_state.clone())
+                .unwrap_or_else(|| "current behavior, health, loop state".to_string()),
+        ),
+        OperatorSection::Behaviors => (
+            "Behaviors",
+            format!(
+                "{} profiles",
+                store
+                    .behaviors
+                    .iter()
+                    .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                    .count()
+            ),
+        ),
+        OperatorSection::Backends => (
+            "Backends",
+            format!(
+                "{} inference backends",
+                backend_ids_for_agent(store, selected_agent_did).len()
+            ),
+        ),
+        OperatorSection::ToolSelections => (
+            "Tool selections",
+            format!(
+                "{} presets",
+                store
+                    .tool_selections
+                    .iter()
+                    .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                    .count()
+            ),
+        ),
+        OperatorSection::InferenceProfiles => (
+            "Inference profiles",
+            format!(
+                "{} profiles",
+                inference_profile_ids_for_agent(store, selected_agent_did).len()
+            ),
+        ),
+        OperatorSection::ScheduledTasks => (
+            "Scheduled Tasks",
+            format!(
+                "{} tasks",
+                store
+                    .scheduled_tasks
+                    .iter()
+                    .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                    .count()
+            ),
+        ),
+        OperatorSection::RequestTimeline => (
+            "Request Timeline",
+            format!(
+                "{} requests",
+                store
+                    .requests
+                    .iter()
+                    .filter(|row| row.agent_did.as_deref() == selected_agent_did)
+                    .count()
+            ),
+        ),
+        OperatorSection::RecentFailures => (
+            "Recent Failures",
+            format!(
+                "{} failures",
+                recent_failure_summaries(store, selected_agent_did).len()
+            ),
+        ),
+    }
+}
+
+pub(super) fn draft_matches_selection(
+    draft: &Option<OperatorDraft>,
+    draft_source_entity_id: Option<&str>,
+    section: OperatorSection,
+    selected_entity_id: Option<&str>,
+) -> bool {
+    match (draft, draft_source_entity_id, selected_entity_id) {
+        (Some(draft), Some(source_entity_id), Some(entity_id)) => {
+            draft.section() == section && source_entity_id == entity_id
+        }
+        (None, _, None) => true,
+        _ => false,
+    }
+}
+
+pub(super) fn draft_for_selection(
+    store: &ClientStore,
+    section: OperatorSection,
+    selected_agent_did: Option<&str>,
+    entity_id: &str,
+) -> Option<OperatorDraft> {
+    match section {
+        OperatorSection::Behaviors => store
+            .behaviors
+            .iter()
+            .find(|row| {
+                row.behavior_id == entity_id && row.agent_did.as_deref() == selected_agent_did
+            })
+            .map(|row| {
+                OperatorDraft::Behavior(BehaviorDraft {
+                    behavior_id: row.behavior_id.clone(),
+                    agent_did: row.agent_did.clone().unwrap_or_default(),
+                    display_name: row.display_name.clone().unwrap_or_default(),
+                    system_prompt: row.system_prompt.clone().unwrap_or_default(),
+                    backend_id: row.backend_id.clone().unwrap_or_default(),
+                    model_name: row.model_name.clone().unwrap_or_default(),
+                    tool_selection_id: row.tool_selection_id.clone().unwrap_or_default(),
+                    inference_profile_id: row.inference_profile_id.clone().unwrap_or_default(),
+                    compaction_strategy: row.compaction_strategy.clone().unwrap_or_default(),
+                    compaction_threshold: row
+                        .compaction_threshold
+                        .map(|value| value.to_string())
+                        .unwrap_or_default(),
+                    enabled: row.enabled.unwrap_or(true),
+                    created_at: row.created_at.clone().unwrap_or_default(),
+                })
+            }),
+        OperatorSection::Backends => {
+            let backend_ids = backend_ids_for_agent(store, selected_agent_did);
+            store
+                .inference_backends
+                .iter()
+                .find(|row| {
+                    row.backend_id == entity_id && backend_ids.contains(&row.backend_id.as_str())
+                })
+                .map(|row| {
+                    OperatorDraft::Backend(BackendDraft {
+                        backend_id: row.backend_id.clone(),
+                        name: row.name.clone().unwrap_or_default(),
+                        provider_kind: row.provider_kind.clone().unwrap_or_default(),
+                        endpoint: row.endpoint.clone().unwrap_or_default(),
+                        api_key: row.api_key.clone().unwrap_or_default(),
+                        api_key_env_var: row.api_key_env_var.clone().unwrap_or_default(),
+                        max_concurrent: row
+                            .max_concurrent
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        max_queue_depth: row
+                            .max_queue_depth
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        enabled: row.enabled.unwrap_or(true),
+                        models: row.models.join(", "),
+                        probe_status: row.probe_status.clone().unwrap_or_default(),
+                    })
+                })
+        }
+        OperatorSection::ToolSelections => store
+            .tool_selections
+            .iter()
+            .find(|row| {
+                row.selection_id == entity_id && row.agent_did.as_deref() == selected_agent_did
+            })
+            .map(|row| {
+                OperatorDraft::ToolSelection(ToolSelectionDraft {
+                    selection_id: row.selection_id.clone(),
+                    agent_did: row.agent_did.clone().unwrap_or_default(),
+                    display_name: row.display_name.clone().unwrap_or_default(),
+                    enable_file_tools: row.enable_file_tools.unwrap_or(false),
+                    file_tools_mode: row.file_tools_mode.clone().unwrap_or_default(),
+                    enable_bash: row.enable_bash.unwrap_or(false),
+                    bash_mode: row.bash_mode.clone().unwrap_or_default(),
+                    cli_tool_names: row.cli_tool_names.join(", "),
+                    enable_meta_tools: row.enable_meta_tools.unwrap_or(false),
+                    delegate_to: row.delegate_to.join(", "),
+                })
+            }),
+        OperatorSection::InferenceProfiles => {
+            let profile_ids = inference_profile_ids_for_agent(store, selected_agent_did);
+            store
+                .inference_profiles
+                .iter()
+                .find(|row| {
+                    row.profile_id == entity_id && profile_ids.contains(&row.profile_id.as_str())
+                })
+                .map(|row| {
+                    OperatorDraft::InferenceProfile(InferenceProfileDraft {
+                        profile_id: row.profile_id.clone(),
+                        display_name: row.display_name.clone().unwrap_or_default(),
+                        context_window: row
+                            .context_window
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        max_output_tokens: row
+                            .max_output_tokens
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        max_turns: row
+                            .max_turns
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        temperature: row
+                            .temperature
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        stream_batch_ms: row
+                            .stream_batch_ms
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                        deadline_duration_secs: row
+                            .deadline_duration_secs
+                            .map(|value| value.to_string())
+                            .unwrap_or_default(),
+                    })
+                })
+        }
+        OperatorSection::ScheduledTasks => store
+            .scheduled_tasks
+            .iter()
+            .find(|row| row.task_id == entity_id && row.agent_did.as_deref() == selected_agent_did)
+            .map(|row| {
+                OperatorDraft::ScheduledTask(ScheduledTaskDraft {
+                    task_id: row.task_id.clone(),
+                    agent_did: row.agent_did.clone().unwrap_or_default(),
+                    behavior_id: row.behavior_id.clone().unwrap_or_default(),
+                    name: row.name.clone().unwrap_or_default(),
+                    prompt: row.prompt.clone().unwrap_or_default(),
+                    interval_secs: row
+                        .interval_secs
+                        .map(|value| value.to_string())
+                        .unwrap_or_default(),
+                    enabled: row.enabled.unwrap_or(true),
+                    next_run_at: row.next_run_at.clone().unwrap_or_default(),
+                    last_run_at: row.last_run_at.clone().unwrap_or_default(),
+                    last_status: row.last_status.clone().unwrap_or_default(),
+                    last_error: row.last_error.clone().unwrap_or_default(),
+                    run_count: row
+                        .run_count
+                        .map(|value| value.to_string())
+                        .unwrap_or_default(),
+                    created_at: row.created_at.clone().unwrap_or_default(),
+                    updated_at: row.updated_at.clone().unwrap_or_default(),
+                })
+            }),
+        _ => None,
+    }
+}
